-removed some unused files
Imported-from: https://subversion.assembla.com/svn/AvP/branches/workbranch/trunk@171

+++ 3dc/avp/win95/Frontend/AvP_Menus.h
@@ -286,7 +286,7 @@ typedef struct
 
 	enum AVPMENU_FONT_ID FontToUse;
 
-	int					PositionInTextField;
+	size_t				PositionInTextField;
 
 	int 				WidthLeftForText; //thing for checking whether user input will fit on screen
